Refactoring; separate file for kernels
