+++ src/ebfm/core/LOOP_SNOW.py
@@ -3,9 +3,10 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 import numpy as np
-import math
 
+from .compute_backend import get_backend, ComputeBackend
 from .constants import SECONDS_PER_HOUR
+from .LOOP_SNOW_kernels import _compaction_kernel, _heat_conduction_kernel, _percolation_kernel
 
 # line_profiler support: `profile` is injected as a builtin by kernprof.
 # When running normally, fall back to a no-op so the decorator stays in place.
@@ -13,439 +14,6 @@ try:
     profile  # noqa: F821
 except NameError:
     profile = lambda f: f  # noqa: E731
-
-# ---------------------------------------------------------------------------
-# Optional Numba support
-# Install via:  pip install "ebfm[performance]"
-# Without numba the module falls back to the NumPy path automatically.
-# ---------------------------------------------------------------------------
-try:
-    from numba import njit, prange  # noqa: F401
-
-    _NUMBA_AVAILABLE = True
-except ImportError:
-    _NUMBA_AVAILABLE = False
-
-    def njit(*args, **kwargs):  # noqa: E302
-        """No-op replacement for numba.njit when numba is not installed."""
-
-        def _wrap(fn):
-            return fn
-
-        # Handle both @njit and @njit(parallel=True, ...) call styles
-        return _wrap if kwargs or (args and not callable(args[0])) else args[0]
-
-    prange = range  # type: ignore[assignment]  # noqa: F811
-
-# Dispatch flag, false by default (opt-in via --with-numba).
-# Set to True by main.py when --with-numba is passed and numba is properly available.
-_USE_NUMBA = False
-
-
-# Kernels for compaction, heatconduction and percolation steps
-# Compiled with numba when available and _USE_NUMBA is True.
-@njit(parallel=True, cache=True)
-def _compaction_kernel(
-    subD,  # (gpsum, nl) Output array, updated in-place
-    subZ,  # (gpsum, nl) Output array, updated in-place
-    subT,  # (gpsum, nl)
-    subW,  # (gpsum, nl), Output array
-    subTmean,  # (gpsum, nl) Output array, updated in-place
-    subD_old,  # (gpsum, nl)
-    subZ_old,  # (gpsum, nl)
-    logyearsnow,  # (gpsum, nl)
-    yearsnow,  # (gpsum, nl)
-    WS,  # (gpsum,)
-    Dens_destr_metam,  # (gpsum, nl)
-    Dens_overb_pres,  # (gpsum, nl)
-    Dens_drift,  # (gpsum, nl)
-    surfH,  # (gpsum,), Output array
-    sumWinit,  # (gpsum,)
-    runoff_irr,  # (gpsum,), Output array
-    dt_yearfrac,
-    dt_seconds,
-    dt,
-    Dice,
-    Dfirn,
-    Dwater,
-    g,
-    T0,  # scalar
-    rd,  # scalar
-    Ec,  # scalar
-    Eg,  # scalar
-    dayseconds,
-    tau_drift,
-    compaction_mode,
-):
-    """Per-column compaction kernel, parallelized over gpsum
-
-    1. Firn compaction
-    2. Seasonal snow compaction
-      2.1 Densification by destructive metamorphism
-      2.2 Densification by Overburden pressure
-      2.3 Drifting-snow densification
-    3. Layer-thickness and surface height, adjustment and subW clipping to mliqmax.
-    """
-    gpsum, nl = subD.shape
-    for i in prange(gpsum):
-        # ------ 1. FIRN COMPACTION ------ #
-        for k in range(nl):
-            subTmean[i, k] = subTmean[i, k] * (1.0 - dt_yearfrac) + dt_yearfrac * subT[i, k]
-            cond_firn_k = (compaction_mode == 0) or (subD[i, k] >= Dfirn)
-            if cond_firn_k:
-                if subD[i, k] < 550.0:
-                    grav_const = 0.07 * max(1.435 - 0.151 * logyearsnow[i, k], 0.25)
-                else:
-                    grav_const = 0.03 * max(2.366 - 0.293 * logyearsnow[i, k], 0.25)
-                temp_factor = math.exp(-Ec / (rd * subT[i, k]) + Eg / (rd * subTmean[i, k]))
-                firn_inc = dt_yearfrac * grav_const * yearsnow[i, k] * g * (Dice - subD[i, k]) * temp_factor
-                subD[i, k] += firn_inc
-
-        # ------ 2. SEASONAL SNOW COMPACTION ------ #
-        if compaction_mode == 1:  # firn+snow
-            # ------ 2.1 DENSIFICATION BY DESTRUCTIVE METAMORPHISM ------ #                                  #
-            # Capture pre-DM snow mask (subD < Dfirn) BEFORE modifying subD.
-            # Reuses it for overburden and drifting
-            was_snow = np.empty(nl, dtype=np.bool_)
-            for k in range(nl):
-                was_snow[k] = subD[i, k] < Dfirn
-
-            for k in range(nl):
-                if was_snow[k]:
-                    cc1 = math.exp(-0.046 * max(subD[i, k] - 175.0, 0.0))
-                    cc2 = 1.0 + (1.0 if subW[i, k] != 0.0 else 0.0)
-                    temp_exp = math.exp(0.04 * (subT[i, k] - T0))
-                    snow_inc = cc1 * cc2 * 2.777e-6 * temp_exp * dt_seconds * subD[i, k]
-                    subD[i, k] = min(subD[i, k] + snow_inc, Dice)
-                    Dens_destr_metam[i, k] = snow_inc
-                else:
-                    Dens_destr_metam[i, k] = 0.0
-
-            # ------ 2.2 DENSIFICATION BY OVERBURDEN PRESSURE ------ #
-            # Two passes:
-            # (1) compute Psload from post-DM densities
-            # (2) apply updates
-            # Refactoring into a loop introduces slight rounding errors!
-            psload = np.empty(nl)
-            psload[0] = 0.5 * subD[i, 0] * subZ[i, 0] * g
-            for k in range(1, nl):
-                xm = subD[i, k - 1] * subZ[i, k - 1] * g
-                xk = subD[i, k] * subZ[i, k] * g
-                psload[k] = psload[k - 1] + 0.5 * (xm + xk)
-
-            for k in range(nl):
-                Dens_overb_pres[i, k] = 0.0
-                if was_snow[k]:  # use pre-DM mask to match NumPy cond_snow
-                    cc7 = 4.0 * 7.62237e6 / 250.0 * subD[i, k] / (1.0 + 60.0 * subW[i, k] / (Dwater * subZ[i, k]))
-                    visc = cc7 * math.exp(0.1 * (T0 - subT[i, k]) + 0.023 * subD[i, k])
-                    overb_inc = dt * dayseconds * subD[i, k] * psload[k] / visc
-                    subD[i, k] = min(subD[i, k] + overb_inc, Dice)
-                    Dens_overb_pres[i, k] = dt * dayseconds * subD[i, k] * psload[k] / visc
-
-            # ------ 2.3 DRIFTING SNOW DENSIFICATION ------ #
-            # Refactoring into a loop introduces slight rounding errors!
-            # Use updated subD < Dfirn to match cond_drift_total
-            z_i_k = 0.0  # z_i[0] = 0
-            for k in range(nl):
-                d_k = max(subD[i, k], 50.0)
-                mo_k = -0.069 + 0.66 * (1.25 - 0.0042 * (d_k - 50.0))
-                si_k = -2.868 * math.exp(-0.085 * WS[i]) + 1.0 + mo_k
-                gamma_k = max(0.0, si_k * math.exp(-z_i_k / 0.1))
-                Dens_drift[i, k] = 0.0
-                if si_k > 0.0 and subD[i, k] < Dfirn:  # use updated mask to match NumPy cond_drift_total
-                    tau_i_k = tau_drift / gamma_k  # gamma_k > 0 since si_k > 0
-                    drift_inc = dt_seconds * max(350.0 - subD[i, k], 0.0) / tau_i_k
-                    subD[i, k] = min(subD[i, k] + drift_inc, Dice)
-                    Dens_drift[i, k] = drift_inc
-                # z_i[k+1] = z_i[k] + subZ[k] * (3.25 - SI[k])
-                z_i_k += subZ[i, k] * (3.25 - si_k)
-
-        # ------ 3. UPDATE LAYER THICKNESS & SURFACE HEIGHT AFTER COMPACTION ------
-        z_sum = 0.0
-        z_sum_old = 0.0
-        subW_sum = 0.0
-        for k in range(nl):
-            if subD[i, k] < Dice:
-                subZ[i, k] = subZ_old[i, k] * subD_old[i, k] / subD[i, k]
-                exp_f = 0.0143 * math.exp(3.3 * (Dice - subD[i, k]) / Dice)
-                denom = 1.0 - exp_f
-                mliqmax_k = subD[i, k] * subZ[i, k] * exp_f / denom * 0.05 * min(Dice - subD[i, k], 20.0)
-                if subW[i, k] > mliqmax_k:
-                    subW[i, k] = mliqmax_k
-            else:
-                # Ice layer: mliqmax = 0 => clamp subW to zero
-                subW[i, k] = 0.0
-            z_sum += subZ[i, k]
-            z_sum_old += subZ_old[i, k]
-            subW_sum += subW[i, k]
-
-        surfH[i] += z_sum - z_sum_old
-        runoff_irr[i] = sumWinit[i] - subW_sum
-
-
-@njit(parallel=True, cache=True)
-def _heat_conduction_kernel(
-    subT,  # (gpsum, nl) Output array, updated in-place
-    Tsurf,  # (gpsum,)
-    kk_sz_top,  # (gpsum,)
-    kk_sz_mid,  # (gpsum, nl-2)
-    dz1,  # (gpsum,)
-    dz2,  # (gpsum, nl-2)
-    denom_l1,  # (gpsum,)
-    denom_interior,  # (gpsum, nl-3)
-    denom_bottom,  # (gpsum,)
-    dt_stab,  # (gpsum,)
-    dt,  # scalar — total time step in days
-    dayseconds,  # scalar — seconds per day
-    geothermal_flux,  # scalar — W m-2
-):
-    """Per-column heat-conduction kernel, parallelized over gpsum.
-
-    Each column is independent over gpsum: the CFL sub-stepping while-loop runs entirely
-    per grid point with no inter-column communication,
-
-    For each sub-step inside the while-loop:
-      1. Compute all inter-layer heat fluxes ``kdTdz`` from the *current*
-         column temperatures ``T_loc``.
-      2. Update ``T_loc`` in-place using ``kdTdz``.
-      3. Repeat until the full time step ``dt`` is covered.
-    """
-    gpsum, nl = subT.shape
-    for i in prange(gpsum):
-        # Thread-local working copy and flux array
-        T_loc = subT[i, :].copy()
-        kdTdz = np.zeros(nl)
-
-        tt_i = 0.0
-        while tt_i < dt:
-            dt_temp_i = min(dt_stab[i], dt - tt_i)
-            if dt_temp_i == 0.0:
-                break
-            tt_i += dt_temp_i
-            C_day_dt = dayseconds * dt_temp_i
-
-            # ---- Step 1: freeze all fluxes from current T_loc ----
-            kdTdz[1] = kk_sz_top[i] * (T_loc[1] - Tsurf[i]) / dz1[i]
-            for k in range(2, nl):
-                kdTdz[k] = kk_sz_mid[i, k - 2] * (T_loc[k] - T_loc[k - 1]) / dz2[i, k - 2]
-
-            # ---- Step 2: update T_loc in-place (kdTdz is now frozen) ----
-            T_loc[1] += C_day_dt * (kdTdz[2] - kdTdz[1]) / denom_l1[i]
-            for k in range(2, nl - 1):
-                T_loc[k] += C_day_dt * (kdTdz[k + 1] - kdTdz[k]) / denom_interior[i, k - 2]
-            T_loc[nl - 1] += C_day_dt * (geothermal_flux - kdTdz[nl - 1]) / denom_bottom[i]
-
-        # Write results back to the shared subT array
-        for k in range(nl):
-            subT[i, k] = T_loc[k]
-
-
-@njit(parallel=True, cache=True)
-def _percolation_kernel(
-    subT,  # (gpsum, nl) Output array, updated in-place
-    subD,  # (gpsum, nl) Output array, updated in-place
-    subW,  # (gpsum, nl) Output array, updated in-place
-    subS,  # (gpsum, nl) Output array, rewritten
-    subZ,  # (gpsum, nl)
-    subW_old,  # (gpsum, nl)
-    avail_W,  # (gpsum,)
-    RP,  # (gpsum, nl)
-    runoff_surface,  # (gpsum,)
-    runoff_slush,  # (gpsum,)
-    refr_P,  # (gpsum,)
-    refr_S,  # (gpsum,)
-    refr_I,  # (gpsum,)
-    slushw,  # (gpsum,)
-    irrw,  # (gpsum,)
-    T0,
-    Dice,
-    Dwater,
-    Lm,
-    Trunoff,
-    perc_depth,
-    percolation_mode,  # 0=bucket, 1=normal, 2=linear, 3=uniform
-    dt,
-):
-    """Per-column percolation, slush storage and refreezing kernel, parallelized over gpsum
-
-    1. Compute Wlim and Wirr (refreezing potential and available irreducible water storate)
-    2. carrot distribution profile (bucket / normal / linear / uniform)
-    3. Refreezing and irreducible-water-storage
-    4. Slush storage
-    5. Slush refreezing
-    6. Irreducible-water refreezing
-    """
-    gpsum, nl = subT.shape
-    sigma2_2 = 2.0 * (perc_depth / 3.0) ** 2
-    norm_coeff = 2.0 / (perc_depth / 3.0) / math.sqrt(2.0 * math.pi)
-    trunoff_factor = 1.0 / (1.0 + dt / Trunoff)
-
-    for i in prange(gpsum):
-        # ------ Refreezing and Irreducible Water Storage Limits ------
-        # Compute refreezing potential (`Wlim`) per layer
-        # Compute maximum irreducible water storage (`mliqmax`)
-        # Compute available irreducible water storage (`Wirr`) per layer
-        wlim_loc = np.empty(nl)
-        wirr_loc = np.empty(nl)
-        for k in range(nl):
-            cpi_k = 152.2 + 7.122 * subT[i, k]
-            c1_k = cpi_k * subD[i, k] * subZ[i, k] * (T0 - subT[i, k]) / Lm
-            c2_k = subZ[i, k] * (1.0 - subD[i, k] / Dice) * Dice
-            wlim_loc[k] = max(min(c1_k, c2_k), 0.0)
-            if subD[i, k] < Dice - 1.0:
-                factor_k = 3.3 * (Dice - subD[i, k]) / Dice
-                exp_f = math.exp(factor_k)
-                irr_f = 0.0143 * exp_f / (1.0 - 0.0143 * exp_f)
-                mliqmax_k = subD[i, k] * subZ[i, k] * irr_f * 0.05 * min(Dice - subD[i, k], 20.0)
-            else:
-                mliqmax_k = 0.0
-            wirr_loc[k] = mliqmax_k - subW_old[i, k]
-
-        # ------ Compute carrot (water-distribution profile) by percolation mode ------
-        carrot_loc = np.zeros(nl)
-        if percolation_mode == 0:  # bucket: all water enters surface layer
-            carrot_loc[0] = 1.0
-        else:
-            # Compute zz (midpoint depth of each layer) for mode 1/2/3
-            depth = 0.0
-            for k in range(nl):
-                zz_k = depth + 0.5 * subZ[i, k]
-                if percolation_mode == 1:  # normal (Gaussian)
-                    carrot_loc[k] = norm_coeff * math.exp(-(zz_k * zz_k) / sigma2_2)
-                elif percolation_mode == 2:  # linear
-                    v = 2.0 * (perc_depth - zz_k) / (perc_depth * perc_depth)
-                    carrot_loc[k] = v if v > 0.0 else 0.0
-                else:  # uniform (mode 3): temporarily store zz for the argmin pass below
-                    carrot_loc[k] = zz_k
-                depth += subZ[i, k]
-
-            if percolation_mode == 3:  # uniform: resolve argmin, then fill layers 0..ind
-                min_dist = math.inf
-                ind = 0
-                for k in range(nl):
-                    d = abs(carrot_loc[k] - perc_depth)
-                    if d < min_dist:
-                        min_dist = d
-                        ind = k
-                for k in range(nl):
-                    carrot_loc[k] = (1.0 / perc_depth) if k <= ind else 0.0
-
-        # Scale by layer thickness, normalize, multiply by avail_W
-        s = 0.0
-        for k in range(nl):
-            carrot_loc[k] *= subZ[i, k]
-            s += carrot_loc[k]
-        avail_W_i = avail_W[i]
-        for k in range(nl):
-            carrot_loc[k] = carrot_loc[k] / s * avail_W_i
-
-        #########################################################
-        # Percolation loop: top-to-bottom refreezing + irreducible storage
-        # avail_W_loc carries unabsorbed water forward across layers
-        #########################################################
-        avail_W_loc = 0.0
-        rp_sum = 0.0
-        for n in range(nl):
-            avail_W_loc += carrot_loc[n]
-            rp_n = min(avail_W_loc, wlim_loc[n])
-            RP[i, n] = rp_n
-            excess = avail_W_loc - wlim_loc[n]
-            if excess < 0.0:
-                excess = 0.0
-            new_subW_n = subW_old[i, n] + min(excess, wirr_loc[n])
-            subW[i, n] = new_subW_n
-            avail_W_loc -= rp_n + (new_subW_n - subW_old[i, n])
-            # Temperature and density update after percolating-water refreezing
-            cpi_n = 152.2 + 7.122 * subT[i, n]
-            subT[i, n] += Lm * rp_n / (subD[i, n] * cpi_n * subZ[i, n])
-            subD[i, n] += rp_n / subZ[i, n]
-            rp_sum += rp_n
-
-        avail_W[i] = avail_W_loc  # write leftover back (becomes avail_W for slush section)
-
-        #########################################################
-        # Slush water storage
-        #########################################################
-        slushspace = np.empty(nl)
-        total_slushspace = 0.0
-        for k in range(nl):
-            ss_k = subZ[i, k] * (1.0 - subD[i, k] / Dice) * Dwater - subW[i, k]
-            if ss_k < 0.0:
-                ss_k = 0.0
-            slushspace[k] = ss_k
-            total_slushspace += ss_k
-
-        # Old slush + leftover water form the available slush input
-        old_slush_sum = 0.0
-        for k in range(nl):
-            old_slush_sum += subS[i, k]
-        avail_W_slush = avail_W_loc + old_slush_sum
-
-        surf_ro = avail_W_slush - total_slushspace
-        runoff_surface[i] = surf_ro if surf_ro > 0.0 else 0.0
-        avail_S = avail_W_slush if avail_W_slush < total_slushspace else total_slushspace
-        runoff_slush[i] = avail_S - trunoff_factor * avail_S
-        avail_S = trunoff_factor * avail_S
-        if avail_S < 1e-25:
-            avail_S = 0.0
-
-        # Bottom-up fill of slush pore space
-        for n in range(nl - 1, -1, -1):
-            fill = avail_S if avail_S < slushspace[n] else slushspace[n]
-            subS[i, n] = fill
-            avail_S -= fill
-
-        #####################################
-        # Refreezing of slush water
-        #####################################
-        rs_sum = 0.0
-        for k in range(nl):
-            cpi_k = 152.2 + 7.122 * subT[i, k]
-            c1_k = cpi_k * subD[i, k] * subZ[i, k] * (T0 - subT[i, k]) / Lm
-            c2_k = subZ[i, k] * (1.0 - subD[i, k] / Dice) * Dice
-            wlim_k = min(c1_k, c2_k)
-            rs_k = 0.0
-            if subS[i, k] > 0.0 and subT[i, k] < T0:
-                rs_k = subS[i, k] if subS[i, k] < wlim_k else wlim_k
-                if rs_k < 0.0:
-                    rs_k = 0.0
-            subS[i, k] -= rs_k
-            subT[i, k] += (Lm * rs_k) / (subD[i, k] * cpi_k * subZ[i, k])
-            subD[i, k] += rs_k / subZ[i, k]
-            rs_sum += rs_k
-
-        #########################################################
-        # Irreducible water refreezing
-        #########################################################
-        ri_sum = 0.0
-        for k in range(nl):
-            cpi_k = 152.2 + 7.122 * subT[i, k]
-            c1_k = cpi_k * subD[i, k] * subZ[i, k] * (T0 - subT[i, k]) / Lm
-            c2_k = subZ[i, k] * (1.0 - subD[i, k] / Dice) * Dice
-            wlim_k = min(c1_k, c2_k)
-            ri_k = 0.0
-            if subW[i, k] > 0.0 and subT[i, k] < T0:
-                ri_k = subW[i, k] if subW[i, k] < wlim_k else wlim_k
-                if ri_k < 0.0:
-                    ri_k = 0.0
-            subW[i, k] -= ri_k
-            subT[i, k] += (Lm * ri_k) / (subD[i, k] * cpi_k * subZ[i, k])
-            subD[i, k] += ri_k / subZ[i, k]
-            ri_sum += ri_k
-
-        #########################################################
-        # Scalar outputs per column
-        #########################################################
-        slushw_i = 0.0
-        irrw_i = 0.0
-        for k in range(nl):
-            slushw_i += subS[i, k]
-            irrw_i += subW[i, k]
-        refr_P[i] = 1e-3 * rp_sum
-        refr_S[i] = 1e-3 * rs_sum
-        refr_I[i] = 1e-3 * ri_sum
-        slushw[i] = slushw_i
-        irrw[i] = irrw_i
 
 
 def main(C, OUT, IN, dt, grid, phys):
@@ -664,7 +232,7 @@ def main(C, OUT, IN, dt, grid, phys):
             OUT["runoff_irr"] = np.zeros(_gshape)
 
         # Numba parallel path:
-        if _USE_NUMBA:
+        if get_backend() == ComputeBackend.NUMBA:
             _mode = {"firn_only": 0, "firn+snow": 1}.get(phys["snow_compaction"], -1)
             if _mode < 0:
                 raise ValueError(f"_compaction_kernel: unknown snow_compaction={phys['snow_compaction']!r}")
@@ -700,132 +268,135 @@ def main(C, OUT, IN, dt, grid, phys):
                 48 * 2 * SECONDS_PER_HOUR,
                 _mode,
             )
-            return True
-
-        # Original NumPy path:
-        # ------ FIRN COMPACTION ------ #
-        if phys["snow_compaction"] in ["firn_only", "firn+snow"]:
-            # Pre-compute the logical condition based on the snow compaction type
-            if phys["snow_compaction"] == "firn_only":
-                cond_firn = np.ones_like(OUT["subD"], dtype=bool)  # All values are True in 2D
-            else:  # 'firn+snow'
-                cond_firn = OUT["subD"] >= Dfirn  # Results in the same 2D shape as OUT['subD']
-
-            # Update annual running average subsurface temperature
-            OUT["subTmean"] *= 1 - dt_yearfrac
-            OUT["subTmean"] += dt_yearfrac * OUT["subT"]
-
-            # Set gravitational constants and masks
-            # Use OUT["subD"] and IN["logyearsnow"] directly
-            # cond_firn is already incorporated into both masks
-            grav_const = np.zeros_like(OUT["subD"])
-            low_density_mask = cond_firn & (OUT["subD"] < 550)
-            high_density_mask = cond_firn & (OUT["subD"] >= 550)
-            grav_const[low_density_mask] = 0.07 * np.maximum(1.435 - 0.151 * IN["logyearsnow"][low_density_mask], 0.25)
-            grav_const[high_density_mask] = 0.03 * np.maximum(
-                2.366 - 0.293 * IN["logyearsnow"][high_density_mask], 0.25
-            )
-
-            # Update firn densities
-            temp_factor = np.exp(-C["Ec"] / (C["rd"] * OUT["subT"]) + C["Eg"] / (C["rd"] * OUT["subTmean"]))
-            firn_increment = dt_yearfrac * grav_const * IN["yearsnow"] * C["g"] * (Dice - OUT["subD"]) * temp_factor
-            np.add(OUT["subD"], firn_increment, where=cond_firn, out=OUT["subD"])
         else:
-            raise ValueError("phys.snow_compaction not set correctly!")
+            # NumPy path
+            # ------ FIRN COMPACTION ------ #
+            if phys["snow_compaction"] in ["firn_only", "firn+snow"]:
+                # Pre-compute the logical condition based on the snow compaction type
+                if phys["snow_compaction"] == "firn_only":
+                    cond_firn = np.ones_like(OUT["subD"], dtype=bool)  # All values are True in 2D
+                else:  # 'firn+snow'
+                    cond_firn = OUT["subD"] >= Dfirn  # Results in the same 2D shape as OUT['subD']
 
-        # ------ SEASONAL SNOW COMPACTION ------ #
-        if phys["snow_compaction"] == "firn+snow":
-            # ------ DENSIFICATION BY DESTRUCTIVE METAMORPHISM ------ #
-            # Precompute condition for snow compaction
-            cond_snow = OUT["subD"] < Dfirn
+                # Update annual running average subsurface temperature
+                OUT["subTmean"] *= 1 - dt_yearfrac
+                OUT["subTmean"] += dt_yearfrac * OUT["subT"]
 
-            # Constants for snow compaction
-            CC3, CC4 = 2.777e-6, 0.04
+                # Set gravitational constants and masks
+                # Use OUT["subD"] and IN["logyearsnow"] directly
+                # cond_firn is already incorporated into both masks
+                grav_const = np.zeros_like(OUT["subD"])
+                low_density_mask = cond_firn & (OUT["subD"] < 550)
+                high_density_mask = cond_firn & (OUT["subD"] >= 550)
+                grav_const[low_density_mask] = 0.07 * np.maximum(
+                    1.435 - 0.151 * IN["logyearsnow"][low_density_mask], 0.25
+                )
+                grav_const[high_density_mask] = 0.03 * np.maximum(
+                    2.366 - 0.293 * IN["logyearsnow"][high_density_mask], 0.25
+                )
 
-            # Precompute CC1, CC2, and temp_exp
-            CC1 = np.exp(-0.046 * np.clip(OUT["subD"] - 175, 0, None))
-            CC2 = 1 + (OUT["subW"] != 0)
-            temp_exp = np.exp(CC4 * (OUT["subT"] - C["T0"]))  #
+                # Update firn densities
+                temp_factor = np.exp(-C["Ec"] / (C["rd"] * OUT["subT"]) + C["Eg"] / (C["rd"] * OUT["subTmean"]))
+                firn_increment = dt_yearfrac * grav_const * IN["yearsnow"] * C["g"] * (Dice - OUT["subD"]) * temp_factor
+                np.add(OUT["subD"], firn_increment, where=cond_firn, out=OUT["subD"])
+            else:
+                raise ValueError("phys.snow_compaction not set correctly!")
 
-            # Snow densification increment
-            snow_increment = CC1 * CC2 * CC3 * temp_exp * dt_seconds * OUT["subD"]
+            # ------ SEASONAL SNOW COMPACTION ------ #
+            if phys["snow_compaction"] == "firn+snow":
+                # ------ DENSIFICATION BY DESTRUCTIVE METAMORPHISM ------ #
+                # Precompute condition for snow compaction
+                cond_snow = OUT["subD"] < Dfirn
 
-            # Apply snow increment only to relevant layers
-            np.add(OUT["subD"], snow_increment, where=cond_snow, out=OUT["subD"])
-            np.minimum(OUT["subD"], Dice, where=cond_snow, out=OUT["subD"])
+                # Constants for snow compaction
+                CC3, CC4 = 2.777e-6, 0.04
 
-            # Store densification by destructive metamorphism
-            np.copyto(OUT["Dens_destr_metam"], snow_increment, where=cond_snow)
+                # Precompute CC1, CC2, and temp_exp
+                CC1 = np.exp(-0.046 * np.clip(OUT["subD"] - 175, 0, None))
+                CC2 = 1 + (OUT["subW"] != 0)
+                temp_exp = np.exp(CC4 * (OUT["subT"] - C["T0"]))  #
 
-            # ------ DENSIFICATION BY OVERBURDEN PRESSURE ------ #
-            CC5, CC6 = 0.1, 0.023
-            CC7 = 4.0 * 7.62237e6 / 250.0 * OUT["subD"] * 1 / (1 + 60 * OUT["subW"] * 1 / (C["Dwater"] * OUT["subZ"]))
+                # Snow densification increment
+                snow_increment = CC1 * CC2 * CC3 * temp_exp * dt_seconds * OUT["subD"]
 
-            # Compute load pressure (Psload)
-            OUT_subD_Z = OUT["subD"] * OUT["subZ"] * C["g"]
-            Psload = np.cumsum(OUT_subD_Z, axis=1) - 0.5 * OUT_subD_Z
-            Psload[~cond_snow] = 0
+                # Apply snow increment only to relevant layers
+                np.add(OUT["subD"], snow_increment, where=cond_snow, out=OUT["subD"])
+                np.minimum(OUT["subD"], Dice, where=cond_snow, out=OUT["subD"])
 
-            # Compute viscosity (Visc)
-            temperature_diff = C["T0"] - OUT["subT"]
-            Visc = CC7 * np.exp(CC5 * temperature_diff + CC6 * OUT["subD"])
-            Visc[~cond_snow] = 0
+                # Store densification by destructive metamorphism
+                np.copyto(OUT["Dens_destr_metam"], snow_increment, where=cond_snow)
 
-            # Update densities
-            OUT["subD"][cond_snow] += (
-                dt * C["dayseconds"] * OUT["subD"][cond_snow] * Psload[cond_snow] / Visc[cond_snow]
+                # ------ DENSIFICATION BY OVERBURDEN PRESSURE ------ #
+                CC5, CC6 = 0.1, 0.023
+                CC7 = (
+                    4.0 * 7.62237e6 / 250.0 * OUT["subD"] * 1 / (1 + 60 * OUT["subW"] * 1 / (C["Dwater"] * OUT["subZ"]))
+                )
+
+                # Compute load pressure (Psload)
+                OUT_subD_Z = OUT["subD"] * OUT["subZ"] * C["g"]
+                Psload = np.cumsum(OUT_subD_Z, axis=1) - 0.5 * OUT_subD_Z
+                Psload[~cond_snow] = 0
+
+                # Compute viscosity (Visc)
+                temperature_diff = C["T0"] - OUT["subT"]
+                Visc = CC7 * np.exp(CC5 * temperature_diff + CC6 * OUT["subD"])
+                Visc[~cond_snow] = 0
+
+                # Update densities
+                OUT["subD"][cond_snow] += (
+                    dt * C["dayseconds"] * OUT["subD"][cond_snow] * Psload[cond_snow] / Visc[cond_snow]
+                )
+                np.minimum(OUT["subD"], C["Dice"], where=cond_snow, out=OUT["subD"])
+
+                # Store densification by overburden pressure
+                OUT["Dens_overb_pres"][cond_snow] = (
+                    dt * C["dayseconds"] * OUT["subD"][cond_snow] * Psload[cond_snow] / Visc[cond_snow]
+                )
+
+                # ------ DRIFTING SNOW DENSIFICATION ------ #
+                MO = -0.069 + 0.66 * (1.25 - 0.0042 * (np.maximum(OUT["subD"], 50) - 50))
+                SI = -2.868 * np.exp(-0.085 * IN["WS"][:, np.newaxis]) + 1 + MO
+                cond_drift = SI > 0
+
+                z_i = np.zeros_like(OUT["subZ"])
+                if nl > 1:
+                    z_i[:, 1:] = np.cumsum(OUT["subZ"][:, :-1] * (3.25 - SI[:, :-1]), axis=1)
+                gamma_drift = np.maximum(0, SI * np.exp(-z_i / 0.1))
+                tau = 48 * 2 * SECONDS_PER_HOUR
+                with np.errstate(divide="ignore", invalid="ignore"):
+                    tau_i = tau / gamma_drift
+
+                # Update densities
+                drift_increment = dt_seconds * np.maximum(350 - OUT["subD"], 0) / tau_i
+                cond_drift_total = cond_drift & (OUT["subD"] < Dfirn)
+                np.add(OUT["subD"], drift_increment, where=cond_drift_total, out=OUT["subD"])
+                np.minimum(OUT["subD"], Dice, where=cond_drift_total, out=OUT["subD"])
+
+                # Store densification by wind shearing
+                np.copyto(OUT["Dens_drift"], drift_increment, where=cond_drift_total)
+
+            # ------ UPDATE LAYER THICKNESS & SURFACE HEIGHT AFTER COMPACTION ------ #
+            cond_layers = OUT["subD"] < Dice
+
+            # Update layer thickness
+            OUT["subZ"][cond_layers] = subZ_old[cond_layers] * subD_old[cond_layers] / OUT["subD"][cond_layers]
+
+            # Update irreducible water storage
+            exp_factor = 0.0143 * np.exp(3.3 * (Dice - OUT["subD"][cond_layers]) / Dice)
+            mliqmax[cond_layers] = (
+                OUT["subD"][cond_layers]
+                * OUT["subZ"][cond_layers]
+                * exp_factor
+                / (1 - exp_factor)
+                * 0.05
+                * np.minimum(Dice - OUT["subD"][cond_layers], 20)
             )
-            np.minimum(OUT["subD"], C["Dice"], where=cond_snow, out=OUT["subD"])
+            OUT["subW"] = np.minimum(mliqmax, OUT["subW"])
 
-            # Store densification by overburden pressure
-            OUT["Dens_overb_pres"][cond_snow] = (
-                dt * C["dayseconds"] * OUT["subD"][cond_snow] * Psload[cond_snow] / Visc[cond_snow]
-            )
-
-            # ------ DRIFTING SNOW DENSIFICATION ------ #
-            MO = -0.069 + 0.66 * (1.25 - 0.0042 * (np.maximum(OUT["subD"], 50) - 50))
-            SI = -2.868 * np.exp(-0.085 * IN["WS"][:, np.newaxis]) + 1 + MO
-            cond_drift = SI > 0
-
-            z_i = np.zeros_like(OUT["subZ"])
-            if nl > 1:
-                z_i[:, 1:] = np.cumsum(OUT["subZ"][:, :-1] * (3.25 - SI[:, :-1]), axis=1)
-            gamma_drift = np.maximum(0, SI * np.exp(-z_i / 0.1))
-            tau = 48 * 2 * SECONDS_PER_HOUR
-            with np.errstate(divide="ignore", invalid="ignore"):
-                tau_i = tau / gamma_drift
-
-            # Update densities
-            drift_increment = dt_seconds * np.maximum(350 - OUT["subD"], 0) / tau_i
-            cond_drift_total = cond_drift & (OUT["subD"] < Dfirn)
-            np.add(OUT["subD"], drift_increment, where=cond_drift_total, out=OUT["subD"])
-            np.minimum(OUT["subD"], Dice, where=cond_drift_total, out=OUT["subD"])
-
-            # Store densification by wind shearing
-            np.copyto(OUT["Dens_drift"], drift_increment, where=cond_drift_total)
-
-        # ------ UPDATE LAYER THICKNESS & SURFACE HEIGHT AFTER COMPACTION ------ #
-        cond_layers = OUT["subD"] < Dice
-
-        # Update layer thickness
-        OUT["subZ"][cond_layers] = subZ_old[cond_layers] * subD_old[cond_layers] / OUT["subD"][cond_layers]
-
-        # Update irreducible water storage
-        exp_factor = 0.0143 * np.exp(3.3 * (Dice - OUT["subD"][cond_layers]) / Dice)
-        mliqmax[cond_layers] = (
-            OUT["subD"][cond_layers]
-            * OUT["subZ"][cond_layers]
-            * exp_factor
-            / (1 - exp_factor)
-            * 0.05
-            * np.minimum(Dice - OUT["subD"][cond_layers], 20)
-        )
-        OUT["subW"] = np.minimum(mliqmax, OUT["subW"])
-
-        # Update surface height and runoff
-        shift = np.sum(OUT["subZ"], axis=1) - np.sum(subZ_old, axis=1)
-        OUT["surfH"] += shift
-        OUT["runoff_irr"] = OUT["sumWinit"] - np.sum(OUT["subW"], axis=1)
+            # Update surface height and runoff
+            shift = np.sum(OUT["subZ"], axis=1) - np.sum(subZ_old, axis=1)
+            OUT["surfH"] += shift
+            OUT["runoff_irr"] = OUT["sumWinit"] - np.sum(OUT["subW"], axis=1)
 
         return True
 
@@ -861,7 +432,7 @@ def main(C, OUT, IN, dt, grid, phys):
         denom_bottom = c_eff[:, -1] * (0.25 * OUT["subZ"][:, -2] + 0.75 * OUT["subZ"][:, -1])
 
         # ------ Heat Conduction Loop ------
-        if _USE_NUMBA:
+        if get_backend() == ComputeBackend.NUMBA:
             # Numba parallel path: prange(gpsum), each column solved independently.
             _heat_conduction_kernel(
                 OUT["subT"],
@@ -946,7 +517,7 @@ def main(C, OUT, IN, dt, grid, phys):
         subW_old = OUT["subW"].copy()  # Store the old water content
         gpsum, nl = OUT["subT"].shape
 
-        if _USE_NUMBA:
+        if get_backend() == ComputeBackend.NUMBA:
             # Numba parallel path:
             _p_mode = {"bucket": 0, "normal": 1, "linear": 2, "uniform": 3}.get(phys["percolation"], -1)
             if _p_mode < 0:
@@ -1003,182 +574,185 @@ def main(C, OUT, IN, dt, grid, phys):
             OUT["slushw"] = _slushw
             OUT["irrw"] = _irrw
             OUT["cpi"] = 152.2 + 7.122 * OUT["subT"]
-            return True
-
-        # Original NumPy path:
-        # ------ Water Input ------
-        avail_W = (
-            OUT["melt"] * 1e3  # Meltwater
-            + IN["rain"] * 1e3  # Rainfall
-            + (OUT["moist_condensation"] - OUT["moist_evaporation"]) * 1e3  # Condensation or evaporation
-        )
-        avail_W = np.maximum(avail_W, 0)  # Ensure no negative water availability
-
-        # ------ Refreezing and Irreducible Water Storage Limits ------
-        OUT["cpi"] = 152.2 + 7.122 * OUT["subT"]  # Specific heat capacity
-        c1 = OUT["cpi"] * OUT["subD"] * OUT["subZ"] * (C["T0"] - OUT["subT"]) / C["Lm"]
-        c2 = OUT["subZ"] * (1 - OUT["subD"] / C["Dice"]) * C["Dice"]
-
-        # Compute refreezing potential (`Wlim`) per layer
-        # np.minimum(c1, c2) is equivalent to np.where(c1 >= c2, c2, c1)
-        Wlim = np.maximum(np.minimum(c1, c2), 0)
-
-        # Maximum irreducible water storage (`mliqmax`)
-        mliqmax = np.zeros_like(OUT["subD"])
-        noice = OUT["subD"] < (C["Dice"] - 1)
-        factor = 3.3 * (C["Dice"] - OUT["subD"][noice]) / C["Dice"]
-        exp_factor = np.exp(factor)
-        irr_factor = 0.0143 * exp_factor / (1 - 0.0143 * exp_factor)
-        mliqmax[noice] = (
-            OUT["subD"][noice] * OUT["subZ"][noice] * irr_factor * 0.05 * np.minimum(C["Dice"] - OUT["subD"][noice], 20)
-        )
-
-        # Available irreducible water storage
-        Wirr = mliqmax - subW_old
-
-        # ------ Water Percolation ------
-        z0 = C["perc_depth"]
-        zz = np.cumsum(OUT["subZ"], axis=1) - 0.5 * OUT["subZ"]
-        carrot = np.zeros_like(OUT["subZ"])
-
-        if phys["percolation"] == "bucket":
-            carrot[:, 0] = 1  # All water is added at the surface layer
-        elif phys["percolation"] == "normal":
-            carrot = 2 * np.exp(-(zz**2) / (2 * (z0 / 3) ** 2)) / (z0 / 3) / np.sqrt(2 * np.pi)
-        elif phys["percolation"] == "linear":
-            carrot = 2 * (z0 - zz) / z0**2
-            carrot = np.maximum(carrot, 0)
-        elif phys["percolation"] == "uniform":
-            ind = np.argmin(np.abs(zz - z0), axis=1)
-            carrot[np.arange(carrot.shape[0]), : ind + 1] = 1 / z0
         else:
-            raise ValueError("`phys['percolation']` is not set correctly!")
+            # NumPy path
+            # ------ Water Input ------
+            avail_W = (
+                OUT["melt"] * 1e3  # Meltwater
+                + IN["rain"] * 1e3  # Rainfall
+                + (OUT["moist_condensation"] - OUT["moist_evaporation"]) * 1e3  # Condensation or evaporation
+            )
+            avail_W = np.maximum(avail_W, 0)  # Ensure no negative water availability
 
-        carrot *= OUT["subZ"]  # Scale by layer thickness
-        carrot /= np.sum(carrot, axis=1)[:, np.newaxis]  # Normalize per layer
-        carrot *= avail_W[:, np.newaxis]  # Distribute water input among layers
+            # ------ Refreezing and Irreducible Water Storage Limits ------
+            OUT["cpi"] = 152.2 + 7.122 * OUT["subT"]  # Specific heat capacity
+            c1 = OUT["cpi"] * OUT["subD"] * OUT["subZ"] * (C["T0"] - OUT["subT"]) / C["Lm"]
+            c2 = OUT["subZ"] * (1 - OUT["subD"] / C["Dice"]) * C["Dice"]
 
-        # ------ Refreezing and Irreducible Water Storage Iteration ------
-        # RP reused across calls: same check-and-fill pattern as Dens_* in compaction
-        _rp_shape = OUT["subZ"].shape
-        if "_perc_RP" not in OUT or OUT["_perc_RP"].shape != _rp_shape:
-            OUT["_perc_RP"] = np.zeros(_rp_shape)
-        else:
-            OUT["_perc_RP"].fill(0.0)
-        RP = OUT["_perc_RP"]
+            # Compute refreezing potential (`Wlim`) per layer
+            # np.minimum(c1, c2) is equivalent to np.where(c1 >= c2, c2, c1)
+            Wlim = np.maximum(np.minimum(c1, c2), 0)
 
-        avail_W_loc = np.zeros(grid["gpsum"])  # Available water per layer
+            # Maximum irreducible water storage (`mliqmax`)
+            mliqmax = np.zeros_like(OUT["subD"])
+            noice = OUT["subD"] < (C["Dice"] - 1)
+            factor = 3.3 * (C["Dice"] - OUT["subD"][noice]) / C["Dice"]
+            exp_factor = np.exp(factor)
+            irr_factor = 0.0143 * exp_factor / (1 - 0.0143 * exp_factor)
+            mliqmax[noice] = (
+                OUT["subD"][noice]
+                * OUT["subZ"][noice]
+                * irr_factor
+                * 0.05
+                * np.minimum(C["Dice"] - OUT["subD"][noice], 20)
+            )
 
-        for n in range(nl):
-            # Compute available water per layer
-            avail_W_loc += carrot[:, n]
+            # Available irreducible water storage
+            Wirr = mliqmax - subW_old
 
-            # RP = min(available water, refreezing capacity), correct for both cond1 and ~cond1
-            # No branching: when avail <= Wlim the full amount refreezes; when avail > Wlim
-            # only Wlim refreezes and the excess goes into irreducible storage.
-            np.minimum(avail_W_loc, Wlim[:, n], out=RP[:, n])
-            # Excess water after refreezing, clamped to zero when there is none
-            excess = np.maximum(avail_W_loc - Wlim[:, n], 0.0)
-            OUT["subW"][:, n] = subW_old[:, n] + np.minimum(excess, Wirr[:, n])
+            # ------ Water Percolation ------
+            z0 = C["perc_depth"]
+            zz = np.cumsum(OUT["subZ"], axis=1) - 0.5 * OUT["subZ"]
+            carrot = np.zeros_like(OUT["subZ"])
 
-            # Deduct water consumed (refrozen + stored) from the running total
-            avail_W_loc -= RP[:, n] + (OUT["subW"][:, n] - subW_old[:, n])
+            if phys["percolation"] == "bucket":
+                carrot[:, 0] = 1  # All water is added at the surface layer
+            elif phys["percolation"] == "normal":
+                carrot = 2 * np.exp(-(zz**2) / (2 * (z0 / 3) ** 2)) / (z0 / 3) / np.sqrt(2 * np.pi)
+            elif phys["percolation"] == "linear":
+                carrot = 2 * (z0 - zz) / z0**2
+                carrot = np.maximum(carrot, 0)
+            elif phys["percolation"] == "uniform":
+                ind = np.argmin(np.abs(zz - z0), axis=1)
+                carrot[np.arange(carrot.shape[0]), : ind + 1] = 1 / z0
+            else:
+                raise ValueError("`phys['percolation']` is not set correctly!")
 
-            # Update temperature and density after refreezing
-            refreeze_heat = C["Lm"] * RP[:, n]
-            OUT["subT"][:, n] += refreeze_heat / (OUT["subD"][:, n] * OUT["cpi"][:, n] * OUT["subZ"][:, n])
-            OUT["subD"][:, n] += RP[:, n] / OUT["subZ"][:, n]
+            carrot *= OUT["subZ"]  # Scale by layer thickness
+            carrot /= np.sum(carrot, axis=1)[:, np.newaxis]  # Normalize per layer
+            carrot *= avail_W[:, np.newaxis]  # Distribute water input among layers
 
-        # Update leftover water
-        avail_W = avail_W_loc
+            # ------ Refreezing and Irreducible Water Storage Iteration ------
+            # RP reused across calls: same check-and-fill pattern as Dens_* in compaction
+            _rp_shape = OUT["subZ"].shape
+            if "_perc_RP" not in OUT or OUT["_perc_RP"].shape != _rp_shape:
+                OUT["_perc_RP"] = np.zeros(_rp_shape)
+            else:
+                OUT["_perc_RP"].fill(0.0)
+            RP = OUT["_perc_RP"]
 
-        #########################################################
-        # Slush water storage
-        #########################################################
+            avail_W_loc = np.zeros(grid["gpsum"])  # Available water per layer
 
-        # Calculate available pore space for storing slush water
-        slushspace = np.maximum(
-            OUT["subZ"] * (1 - OUT["subD"] / C["Dice"]) * C["Dwater"] - OUT["subW"], 0.0
-        )  # shape: [grid.gpsum, nl]
+            for n in range(nl):
+                # Compute available water per layer
+                avail_W_loc += carrot[:, n]
 
-        # Total available slush space across all layers
-        total_slushspace = np.sum(slushspace, axis=1)  # shape: [grid.gpsum]
+                # RP = min(available water, refreezing capacity), correct for both cond1 and ~cond1
+                # No branching: when avail <= Wlim the full amount refreezes; when avail > Wlim
+                # only Wlim refreezes and the excess goes into irreducible storage.
+                np.minimum(avail_W_loc, Wlim[:, n], out=RP[:, n])
+                # Excess water after refreezing, clamped to zero when there is none
+                excess = np.maximum(avail_W_loc - Wlim[:, n], 0.0)
+                OUT["subW"][:, n] = subW_old[:, n] + np.minimum(excess, Wirr[:, n])
 
-        # Calculate surface runoff (excess water at the surface)
-        avail_W += np.sum(OUT["subS"], axis=1)  # Update available water with slush from all layers
-        OUT["runoff_surface"] = np.maximum(avail_W - total_slushspace, 0.0)  # Excess water at the surface
+                # Deduct water consumed (refrozen + stored) from the running total
+                avail_W_loc -= RP[:, n] + (OUT["subW"][:, n] - subW_old[:, n])
 
-        # Update slush water content after new water input and runoff
-        avail_S = np.minimum(avail_W, total_slushspace)  # Available slush water for storage
+                # Update temperature and density after refreezing
+                refreeze_heat = C["Lm"] * RP[:, n]
+                OUT["subT"][:, n] += refreeze_heat / (OUT["subD"][:, n] * OUT["cpi"][:, n] * OUT["subZ"][:, n])
+                OUT["subD"][:, n] += RP[:, n] / OUT["subZ"][:, n]
 
-        # Calculate slush water runoff and reduce avail_S accordingly
-        OUT["runoff_slush"] = avail_S - 1.0 / (1.0 + dt / C["Trunoff"]) * avail_S
-        avail_S = 1.0 / (1.0 + dt / C["Trunoff"]) * avail_S
-        avail_S[avail_S < 1e-25] = 0.0  # Set near-zero available slush to zero
+            # Update leftover water
+            avail_W = avail_W_loc
 
-        # Initialize slush water in all layers (reusing arrays)
-        if "subS" not in OUT or OUT["subS"].shape != (gpsum, nl):
-            OUT["subS"] = np.zeros((gpsum, nl))
-        else:
-            OUT["subS"].fill(0.0)
+            #########################################################
+            # Slush water storage
+            #########################################################
 
-        # Bottom-up filling of pore space with slush water
-        # Each layer takes min(available, pore space), no branching needed
-        for n in range(nl - 1, -1, -1):  # Loop from bottom (nl) to top (1)
-            np.minimum(avail_S, slushspace[:, n], out=OUT["subS"][:, n])
-            avail_S -= OUT["subS"][:, n]
+            # Calculate available pore space for storing slush water
+            slushspace = np.maximum(
+                OUT["subZ"] * (1 - OUT["subD"] / C["Dice"]) * C["Dwater"] - OUT["subW"], 0.0
+            )  # shape: [grid.gpsum, nl]
 
-        #####################################
-        # Refreezing of slush water
-        #####################################
-        # Determine whether cold content or density limits the amount of refreezing
-        OUT["cpi"] = 152.2 + 7.122 * OUT["subT"]
-        c1 = OUT["cpi"] * OUT["subD"] * OUT["subZ"] * (C["T0"] - OUT["subT"]) / C["Lm"]
-        c2 = OUT["subZ"] * (1 - OUT["subD"] / C["Dice"]) * C["Dice"]
-        Wlim = np.minimum(c1, c2)
+            # Total available slush space across all layers
+            total_slushspace = np.sum(slushspace, axis=1)  # shape: [grid.gpsum]
 
-        # Refreezing of slush:
-        # RS = min(subS, Wlim) where layer is cold and wet, else 0
-        # No temporaries needed, min() handles both branches; mask zeroes inactive elements
-        layer_cond = (OUT["subS"] > 0) & (OUT["subT"] < C["T0"])
-        RS = np.minimum(OUT["subS"], Wlim)
-        RS[~layer_cond] = 0.0
+            # Calculate surface runoff (excess water at the surface)
+            avail_W += np.sum(OUT["subS"], axis=1)  # Update available water with slush from all layers
+            OUT["runoff_surface"] = np.maximum(avail_W - total_slushspace, 0.0)  # Excess water at the surface
 
-        # Update slush water content
-        OUT["subS"] -= RS
+            # Update slush water content after new water input and runoff
+            avail_S = np.minimum(avail_W, total_slushspace)  # Available slush water for storage
 
-        # Update temperature after refreezing
-        OUT["subT"] += (C["Lm"] * RS) / (OUT["subD"] * OUT["cpi"] * OUT["subZ"])
+            # Calculate slush water runoff and reduce avail_S accordingly
+            OUT["runoff_slush"] = avail_S - 1.0 / (1.0 + dt / C["Trunoff"]) * avail_S
+            avail_S = 1.0 / (1.0 + dt / C["Trunoff"]) * avail_S
+            avail_S[avail_S < 1e-25] = 0.0  # Set near-zero available slush to zero
 
-        # Update density after refreezing
-        OUT["subD"] += RS / OUT["subZ"]
+            # Initialize slush water in all layers (reusing arrays)
+            if "subS" not in OUT or OUT["subS"].shape != (gpsum, nl):
+                OUT["subS"] = np.zeros((gpsum, nl))
+            else:
+                OUT["subS"].fill(0.0)
 
-        ###########################################################
-        # REFREEZING OF IRREDUCIBLE WATER
-        ###########################################################
-        # Determine whether cold content or density limits the amount of refreezing
-        OUT["cpi"] = 152.2 + 7.122 * OUT["subT"]
-        c1 = OUT["cpi"] * OUT["subD"] * OUT["subZ"] * (C["T0"] - OUT["subT"]) / C["Lm"]
-        c2 = OUT["subZ"] * (1 - OUT["subD"] / C["Dice"]) * C["Dice"]
-        Wlim = np.minimum(c1, c2)
+            # Bottom-up filling of pore space with slush water
+            # Each layer takes min(available, pore space), no branching needed
+            for n in range(nl - 1, -1, -1):  # Loop from bottom (nl) to top (1)
+                np.minimum(avail_S, slushspace[:, n], out=OUT["subS"][:, n])
+                avail_S -= OUT["subS"][:, n]
 
-        # Calculate refreezing amounts
-        valid_mask = (OUT["subW"] > 0) & (OUT["subT"] < C["T0"])
-        RI = np.minimum(OUT["subW"], Wlim)
-        RI[~valid_mask] = 0.0
+            #####################################
+            # Refreezing of slush water
+            #####################################
+            # Determine whether cold content or density limits the amount of refreezing
+            OUT["cpi"] = 152.2 + 7.122 * OUT["subT"]
+            c1 = OUT["cpi"] * OUT["subD"] * OUT["subZ"] * (C["T0"] - OUT["subT"]) / C["Lm"]
+            c2 = OUT["subZ"] * (1 - OUT["subD"] / C["Dice"]) * C["Dice"]
+            Wlim = np.minimum(c1, c2)
 
-        # Update water content (subW), temperature (subT), and density (subD)
-        OUT["subW"] -= RI
-        OUT["subT"] += (C["Lm"] * RI) / (OUT["subD"] * OUT["cpi"] * OUT["subZ"])
-        OUT["subD"] += RI / OUT["subZ"]
+            # Refreezing of slush:
+            # RS = min(subS, Wlim) where layer is cold and wet, else 0
+            # No temporaries needed, min() handles both branches; mask zeroes inactive elements
+            layer_cond = (OUT["subS"] > 0) & (OUT["subT"] < C["T0"])
+            RS = np.minimum(OUT["subS"], Wlim)
+            RS[~layer_cond] = 0.0
 
-        # Determine total refreezing and individual contributions, and total slush water and irreducible water
-        OUT["refr"] = 1e-3 * (np.sum(RP, axis=1) + np.sum(RS, axis=1) + np.sum(RI, axis=1))
-        OUT["refr_P"] = 1e-3 * np.sum(RP, axis=1)  # Refreezing of percolating water
-        OUT["refr_S"] = 1e-3 * np.sum(RS, axis=1)  # Refreezing of slush water
-        OUT["refr_I"] = 1e-3 * np.sum(RI, axis=1)  # Refreezing of irreducible water
-        OUT["slushw"] = np.sum(OUT["subS"], axis=1)  # Total stored slush water
-        OUT["irrw"] = np.sum(OUT["subW"], axis=1)  # Total stored irreducible water
+            # Update slush water content
+            OUT["subS"] -= RS
+
+            # Update temperature after refreezing
+            OUT["subT"] += (C["Lm"] * RS) / (OUT["subD"] * OUT["cpi"] * OUT["subZ"])
+
+            # Update density after refreezing
+            OUT["subD"] += RS / OUT["subZ"]
+
+            ###########################################################
+            # REFREEZING OF IRREDUCIBLE WATER
+            ###########################################################
+            # Determine whether cold content or density limits the amount of refreezing
+            OUT["cpi"] = 152.2 + 7.122 * OUT["subT"]
+            c1 = OUT["cpi"] * OUT["subD"] * OUT["subZ"] * (C["T0"] - OUT["subT"]) / C["Lm"]
+            c2 = OUT["subZ"] * (1 - OUT["subD"] / C["Dice"]) * C["Dice"]
+            Wlim = np.minimum(c1, c2)
+
+            # Calculate refreezing amounts
+            valid_mask = (OUT["subW"] > 0) & (OUT["subT"] < C["T0"])
+            RI = np.minimum(OUT["subW"], Wlim)
+            RI[~valid_mask] = 0.0
+
+            # Update water content (subW), temperature (subT), and density (subD)
+            OUT["subW"] -= RI
+            OUT["subT"] += (C["Lm"] * RI) / (OUT["subD"] * OUT["cpi"] * OUT["subZ"])
+            OUT["subD"] += RI / OUT["subZ"]
+
+            # Determine total refreezing and individual contributions, and total slush water and irreducible water
+            OUT["refr"] = 1e-3 * (np.sum(RP, axis=1) + np.sum(RS, axis=1) + np.sum(RI, axis=1))
+            OUT["refr_P"] = 1e-3 * np.sum(RP, axis=1)  # Refreezing of percolating water
+            OUT["refr_S"] = 1e-3 * np.sum(RS, axis=1)  # Refreezing of slush water
+            OUT["refr_I"] = 1e-3 * np.sum(RI, axis=1)  # Refreezing of irreducible water
+            OUT["slushw"] = np.sum(OUT["subS"], axis=1)  # Total stored slush water
+            OUT["irrw"] = np.sum(OUT["subW"], axis=1)  # Total stored irreducible water
 
         return True
